% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_poisson.r
\name{diff_vec}
\alias{diff_vec}
\title{Calculate vector of differences between desired targets and calculated targets}
\usage{
diff_vec(betavec, wh, xmat, targets, dweights = rep(1, length(targets)))
}
\arguments{
\item{betavec}{vector parameters to poisson model, length s * k (see h, s, k definitions below)}

\item{wh}{vector of household total weights, length h}

\item{xmat}{h x k matrix of data for households}

\item{targets}{s x k matrix of desired target values}

\item{dweights}{optional length s * k vector of weights to multiply the vector of targets}
}
\value{
vector of differences of length s * k
}
\description{
\code{targets_mat} calculates vector of differences of length s * k
}
\details{
See (Khitatrakun, Mermin, Francis, 2016)

\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
}
}
\examples{
p <- make_problem(h=10, s=3, k=2)
dw <- get_dweights(p$targets)
betavec <- rep(0, p$s * p$k)
diff_vec(betavec, p$wh, p$xmat, p$targets, dw)
}
