% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ipopt.r
\name{eval_f_xm1sq}
\alias{eval_f_xm1sq}
\title{Evaluate Objective Function for \code{(x - 1)^2}.}
\usage{
eval_f_xm1sq(x, inputs)
}
\arguments{
\item{x}{Numeric vector representing the ratio of new weights to initial
weights, length h.}

\item{inputs}{List created by \code{reweight}. It must have the following elements:
\describe{
\item{iweight}{vector of initial weights, length h}
\item{objscale}{scalar to divide calculations by}
}}
}
\value{
A scalar, the scaled weighted sum of the differences between x and
1.
}
\description{
Evaluate an objective function that calculates the weighted sum of \code{(x - 1)^2} at a specific point \code{x}, where \code{x} is the ratio of new weights to
existing weights. It weights elements in the sum by the initial weights. It
scales the calculation by a scaling factor. Most users will have no need to
call this function, but it can be useful when debugging a problem.
}
