% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ipopt.r
\name{eval_g}
\alias{eval_g}
\title{Evaluate Constraints.}
\usage{
eval_g(x, inputs)
}
\arguments{
\item{x}{Numeric vector representing the ratio of new weights to initial
weights, length h.}

\item{inputs}{List created by \code{reweight}. It must have a data frame named
\code{cc_sparse} (representing a sparse matrix of constraint coefficients), with
at least the following columns:
\describe{
\item{i}{the constraint number}
\item{j}{an index into x}
\item{nzcc}{a nonzero constraint coefficient}
}}
}
\value{
A vector with one vaue per constraint.
}
\description{
Evaluate constraints (targets) at a specific point. That is, evaluate
constraints at a vector x, the ratio of new weights to initial weights. Most
users will have no need to call this function, but it can be useful when
debugging a problem.
}
