% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoweight.r
\name{geoweight}
\alias{geoweight}
\title{Split weights across geographies}
\usage{
geoweight(
  wh,
  xmat,
  targets,
  dweights = get_dweights(targets),
  method = "LM",
  betavec = rep(0, length(targets)),
  maxiter = NULL,
  opts = NULL,
  quiet = FALSE
)
}
\arguments{
\item{wh}{Household weights, 1 per household, numeric vector length h. Each
household's geography weights must sum to its household weight.}

\item{xmat}{Data for households. Matrix with 1 row per household and 1 column
per characteristic (h x k matrix).}

\item{targets}{Targeted values. Matrix with 1 row per geographic area and 1
column per characteristic.}

\item{dweights}{Difference weights: weights to be applied to Weighting factors for targets (h x k matrix).}

\item{method}{optional parameter for approach to use; must be one of
c('LM', 'Broyden', 'Newton'); default is 'LM'}

\item{betavec}{optional vector of initial guess at parameters, length s * k;
default is zero for all}

\item{maxiter}{integer; defaults: Broyden (2000), Newton (200), LM (200)}

\item{opts}{list of options that will update nelsqv or nls.lm options respectively}

\item{quiet}{c(TRUE, FALSE) FALSE is default; TRUE provides newlsqv or nls.lm output}
}
\value{
A list with the following elements:
\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
\item{solver_message}{message from the solver that was used}
\item{etime}{elapsed time}
\item{beta_opt_mat}{s x k matrix of optimal parameters}
\item{whs}{h x s matrix of state weights for each household, computed
using the optimal parameters}
\item{wh}{the input vector of household total weights, length h}
\item{xmat}{matrix of data for households, h x k}
\item{dweights}{optional vector of weighting factors for targets, length s * k}
\item{output}{list of output from the solver that was used}
}
}
\description{
\code{geoweight} calculates state weights for each household in a microdata
file that add up to the household total weight, such that weighted state
totals for selected characteristics hit or come close to desired targets
}
\details{
\code{geoweight} uses the solver \code{\link[nleqslv]{nleqslv}} or the solver
\code{\link[minpack.lm]{nls.lm}} depending on user choice.

The default method, LM, uses \code{nls.lm} as it appears to be the most robust
of the methods, rarely failing and often producing a better optimum than
Broyden or Newton. However, in some circumstances one of the latter may work
better. It is hard to define guidelines for when a particular method will be
better. The Broyden method can be faster or more robust than the Newton method
but generally requires many more iterations than the Newton method, although
iterations will be faster.
}
\examples{
# Example 1: Determine state weights for a simple problem with random data
p <- make_problem(h=10, s=3, k=2)
dw <- get_dweights(p$targets)

res1 <- geoweight(wh = p$wh, xmat = p$xmat, targets = p$targets,
  dweights = dw, quiet=TRUE)

res2 <- geoweight(wh = p$wh, xmat = p$xmat, targets = p$targets,
  dweights = dw, method = 'Newton', quiet=TRUE)

res3 <- geoweight(wh = p$wh, xmat = p$xmat, targets = p$targets,
  dweights = dw, method = 'LM', quiet=TRUE)

res1
res2
res3
c(res1$sse_unweighted, res2$sse_unweighted, res3$sse_unweighted)

# verify that the state weights produce the desired targets
t(res2$whs) \%*\% p$xmat
p$targets
}
