% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight.r
\name{get_cc_sparse}
\alias{get_cc_sparse}
\title{Create Sparse Constraints Coefficients Data Frame.}
\usage{
get_cc_sparse(xmat, target_names, iweights)
}
\arguments{
\item{xmat}{Numeric matrix of unweighted data values in dense
form where each row is a household, each column corresponds to a
constraint, and each cell is a data value.}

\item{target_names}{Character vector with names for each constraint. One element per
column in \code{xmat}.}

\item{iweights}{Numeric vector of initial weights in the microdata file, one per row in \code{xmat}.}
}
\value{
A data frame representing a sparse matrix of constraint
coefficients, with columns:
\describe{
\item{i}{the constraint number}
\item{j}{an index into x}
\item{cname}{constraint name, from \code{target_names}}
\item{nzcc}{the nonzero constraint coefficient - the initial weight
multiplied by value}
\item{iweight}{initial weight for this cell}
\item{value}{data value}
}
}
\description{
Construct a data frame that represents a sparse matrix with nonzero
constraint coefficients. A constraint coefficient is the amount by which a
constraint changes if an element of the vector \code{x} changes by one unit. This
stores only those coefficients that are nonzero, in triplet form, where \code{i}
is an index identifying which constraint is represented, \code{j} is an index
identifying which element of the vector \code{x} is represented, and \code{nzcc} is the
value of the nonzero constraint coefficient.
}
\details{
The data value in \code{xmat} multiplied by the initial weight determines how much
a constraint will change with a unit change in \code{x}, the ratio of the new
weight to the initial weight. For example, suppose household number 7 has
income of $10,000 then. Suppose that this household is in row 7 of \code{xmat} and
that column 3 of \code{xmat} corresponds to income. If this household has an
initial weight of 20, then the constraint coefficient for this cell - how
much changing the \code{x} value for this household will change the constraint for
total income - is 200,000 (20 x 100,000). The constraints coefficient data
frame will have a row where \code{i} (the index for the constraint, not the household)
is 3, \code{j} (the index for the household) is 7, and \code{nzcc} is 200000.
}
