% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_poisson.r
\name{get_delta}
\alias{get_delta}
\title{Calculate household-specific constants}
\usage{
get_delta(wh, beta, xmat)
}
\arguments{
\item{wh}{vector of household total weights, length h (see h, s, k definitions below)}

\item{beta}{s x k matrix of parameters to poisson model}

\item{xmat}{h x k matrix of data for households

\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
}}
}
\value{
A vector of household-specific constants of length h
}
\description{
\code{get_delta} calculates household-specific constants for the poisson model
}
\details{
See (Khitatrakun, Mermin, Francis, 2016, p.5)
}
\examples{
p <- make_problem(h=10, s=3, k=2)
beta <- matrix(0, nrow=p$s, ncol=p$k)
get_delta(p$wh, beta, p$xmat)
}
