% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_poisson.r
\name{get_dweights}
\alias{get_dweights}
\title{Calculate weights to be multiplied by each target so that it hits a goal}
\usage{
get_dweights(targets, goal = 100)
}
\arguments{
\item{targets}{s x k matrix of desired target values}

\item{goal}{optional scalar -- weights multiplied by targets will equal this}
}
\value{
scalar: sum of squared differences
}
\description{
\code{get_dweights} calculates difference weights
}
\details{
\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
}
}
\examples{
p <- make_problem(h=10, s=3, k=2)
dw <- get_dweights(p$targets)
dw * p$targets
}
