% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight.r
\name{get_inputs}
\alias{get_inputs}
\title{Create List of Inputs for Optimization Functions.}
\usage{
get_inputs(iweights, targets, target_names, tol, cc_sparse, xlb, xub)
}
\arguments{
\item{iweights}{Numeric vector of initial weights in the microdata file.}

\item{targets}{Numeric vector of targets.}

\item{target_names}{Character vector of names for targets.}

\item{tol}{Numeric vector of tolerances, one per target.}

\item{cc_sparse}{Data frame with nonzero constraint coefficients in sparse format.}

\item{xlb}{Numeric vector of lower bounds for the \code{x} vector.}

\item{xub}{Numeric vector of upper bounds for the \code{x} vector.}
}
\description{
Most users will never use this function. Can be useful for debugging.
}
