% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_poisson.r
\name{get_weights}
\alias{get_weights}
\title{Calculate state weights for each household}
\usage{
get_weights(beta, delta, xmat)
}
\arguments{
\item{beta}{s x k matrix of parameters to poisson model}

\item{delta}{vector of household-specific constants, length h}

\item{xmat}{h x k matrix of data for households}
}
\value{
h x s matrix of state weights for each household
}
\description{
\code{get_weights} calculates state weights for each household
}
\details{
See (Khitatrakun, Mermin, Francis, 2016, p.4)

\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
}
}
\examples{
p <- make_problem(h=10, s=3, k=2)
beta <- matrix(0, nrow=p$s, ncol=p$k)
delta <- get_delta(p$wh, beta, p$xmat)
get_weights(beta, delta, p$xmat)
}
