% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{make_problem}
\alias{make_problem}
\title{Create Problem of Desired Size}
\usage{
make_problem(h, s, k)
}
\arguments{
\item{h}{number of households to create.}

\item{s}{number of states to create.}

\item{k}{number of characteristics each household will have.}
}
\value{
A list with elements of a problem.
}
\description{
Will have households, states, and household characteristics.
}
\examples{
make_problem(4, 3, 2)
}
