% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microweight-package.r
\docType{package}
\name{microweight-package}
\alias{microweight}
\alias{microweight-package}
\title{microweight: Tools for Weighting Microdata Files}
\description{
To learn more about dplyr, start with the vignettes:
\code{browseVignettes(package = "microweight")}
}
\author{
\strong{Maintainer}: Don Boyd \email{donboyd5@gmail.com}

}
