% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microweight-package.r
\docType{package}
\name{microweight}
\alias{microweight}
\title{microweight}
\description{
To learn more about \code{microweight}, start with the vignettes:
\code{browseVignettes(package = "microweight")}
}
