% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_poisson.r
\name{sse}
\alias{sse}
\title{Sum of squared differences between desired targets and calculated targets}
\usage{
sse(betavec, wh, xmat, targets, dweights = NULL)
}
\arguments{
\item{betavec}{vector parameters to poisson model, length s * k (see h, s, k definitions below)}

\item{wh}{vector of household total weights, length h}

\item{xmat}{h x k matrix of data for households}

\item{targets}{s x k matrix of desired target values}

\item{dweights}{optional length s * k vector of weights to multiply the vector of targets}
}
\value{
scalar: sum of squared differences
}
\description{
\code{sse} calculates sum of squared differences
}
\details{
\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
}
}
\examples{
p <- make_problem(h=10, s=3, k=2)
dw <- get_dweights(p$targets)
betavec <- rep(0, p$s * p$k)
sse(betavec, p$wh, p$xmat, p$targets, dw)
}
