% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_poisson.r
\name{targets_vec}
\alias{targets_vec}
\title{Calculate targets vector}
\usage{
targets_vec(betavec, wh, xmat, s)
}
\arguments{
\item{betavec}{vector parameters to poisson model, length s * k (see h, s, k definitions below)}

\item{wh}{vector of household total weights, length h}

\item{xmat}{h x k matrix of data for households}

\item{s}{number of states (or other geographies or subgroups)}
}
\value{
vector of targets of length s * k
}
\description{
\code{targets_vec} calculates vector of targets of length s * k
}
\details{
See (Khitatrakun, Mermin, Francis, 2016)

\describe{
\item{h}{number of households (or individuals, records, tax returns, etc.)}
\item{s}{number of states (or other geographies or subgroups)}
\item{k}{number of characteristics each household has}
}
}
\examples{
p <- make_problem(h=10, s=3, k=2)
betavec <- rep(0, p$s * p$k)
targets_vec(betavec, p$wh, p$xmat, p$s)
}
