% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{vtom}
\alias{vtom}
\title{Convert Vector of States and Characteristics to Matrix With States as Rows}
\usage{
vtom(v, s)
}
\arguments{
\item{v}{vector to be converted to matrix.}

\item{s}{number of states.}
}
\value{
A matrix with states as rows and characteristics as columns.
}
\description{
Matrix will have same format as data matrix.
}
\examples{
n_states <- 3
n_chars <- 2
set.seed(1)
value_vec <- runif(n_states * n_chars)
vtom(value_vec, n_states)
}
